+++ jetty-core/jetty-util/src/main/java/org/eclipse/jetty/util/StringUtil.java
@@ -14,7 +14,6 @@
 package org.eclipse.jetty.util;
 
 import java.io.UnsupportedEncodingException;
-import java.nio.CharBuffer;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -479,75 +478,6 @@ public class StringUtil
                 return false;
         }
         return true;
-    }
-
-    /**
-     * <p>Get a zero copy subsequence of a {@link String}.</p>
-     * <p>Use of this is method can result in unforeseen GC consequences and can bypass
-     * JVM optimizations available in {@link String#subSequence(int, int)}.  It should only
-     * be used in cases where there is a known benefit: large sub sequence of a larger string with no retained
-     * references to the sub sequence beyond the life time of the string.</p>
-     * @param string The {@link String} to take a subsequence of.
-     * @param offset The offset in characters into the string to start the subsequence
-     * @param length The length in characters of the substring
-     * @return A new {@link CharSequence} containing the subsequence, backed by the passed {@link String}
-     * or the original {@link String} if it is the same.
-     */
-    public static CharSequence subSequence(String string, int offset, int length)
-    {
-        Objects.requireNonNull(string);
-
-        if (offset == 0 && string.length() == length)
-            return string;
-        if (length == 0)
-            return "";
-
-        int end = offset + length;
-        if (offset < 0 || offset > end || end > string.length())
-            throw new IndexOutOfBoundsException("offset and/or length out of range");
-
-        return new CharSequence()
-        {
-            @Override
-            public int length()
-            {
-                return length;
-            }
-
-            @Override
-            public char charAt(int index)
-            {
-                return string.charAt(offset + index);
-            }
-
-            @Override
-            public CharSequence subSequence(int start, int end)
-            {
-                return StringUtil.subSequence(string, offset + start, end - start);
-            }
-
-            @Override
-            public String toString()
-            {
-                return string.substring(offset, offset + length);
-            }
-        };
-    }
-
-    /**
-     * Get a zero copy subsequence of a {@code char} array.
-     * @param chars The characters to take a subsequence of.  These character are not copied and the array should not be
-     * modified for the life of the returned CharSequence.
-     * @param offset The offset in characters into the string to start the subsequence
-     * @param length The length in characters of the substring
-     * @return A new {@link CharSequence} containing the subsequence.
-     */
-    public static CharSequence subSequence(char[] chars, int offset, int length)
-    {
-        // Needed to make bounds check of wrap the same as for string.substring
-        if (length == 0)
-            return "";
-        return CharBuffer.wrap(chars, offset, length);
     }
 
     public static String toUTF8String(byte[] b, int offset, int length)
